fixup! Extend gather_summary_obs with localization
